+++ Products/CMFPlone/__init__.py
@@ -115,6 +115,9 @@ def initialize(context):
     # Make cgi.escape available TTW
     ModuleSecurityInfo('cgi').declarePublic('escape')
 
+    # Apply monkey patches
+    from Products.CMFPlone import patches  # noqa
+
     # Register unicode splitter w/ ZCTextIndex
     # pipeline registry
     from Products.CMFPlone import UnicodeSplitter  # noqa
@@ -208,6 +211,6 @@ PloneMessageFactory = MessageFactory('plone')
 from zope.i18nmessageid import MessageFactory
 PloneLocalesMessageFactory = MessageFactory('plonelocales')
 
-# Apply monkey patches.  If we do this in the initialize method,
-# it is too late for some of them.
-from Products.CMFPlone import patches  # noqa
+# Apply early monkey patches.  For these patches, it is too late if we do this
+# in the initialize method.
+from Products.CMFPlone import earlypatches  # noqa