+++ Products/CMFPlone/patches/__init__.py
@@ -18,8 +18,6 @@ import speed                    # Various caching patches to improve speed
 import iso8601                  # use `DateTime.ISO8601` for `DateTime.ISO`
 iso8601.applyPatches()
 
-import security					# misc security fixes
-
 import sendmail
 sendmail.applyPatches()